Fix PHPStan cast.double: give min-price helper a typed array shape
The CI run on this branch failed PHPStan (L10) at
novoton_booking/add_to_cart.php:222 with "Cannot cast mixed to float".

Root cause: fn_novoton_min_price_from_xml() already documented its exact
return shape in prose — ['price'=>float, 'room'=>string, 'board'=>string]
— but declared the machine-readable type as array<string, mixed>|null, so
PHPStan saw the 'price' element as mixed and flagged the (float) cast that
commit be76fcc introduced at the call site.

Promote the prose into a real array{price: float, room: string,
board: string}|null shape. The value is provably float (built at
helpers.php:422), so this is truthful, not a suppression. Both call sites
(add_to_cart.php and ajax_recalculate_price.php) are now cast-free.

Verified locally: PHPStan, Psalm, PHPCS, PHP CS Fixer, Rector and php -l
all pass.

+++ addon-novoton-holidays/app/addons/novoton_holidays/controllers/frontend/novoton_booking/add_to_cart.php
@@ -219,7 +219,7 @@ use Tygh\Addons\TravelCore\TravelConstants;
             $room_id_for_match  = rawurldecode(TypeCoerce::toString($bookingData['room_id']));
             $board_id_for_match = TypeCoerce::toString($bookingData['board_id'] ?? '');
             $minPriceMatch = fn_novoton_min_price_from_xml($priceData, $room_id_for_match, $board_id_for_match);
-            $rawPrice = $minPriceMatch !== null ? (float)$minPriceMatch['price'] : (float)((string)$priceData->Price);
+            $rawPrice = $minPriceMatch !== null ? $minPriceMatch['price'] : (float)((string)$priceData->Price);
             $base_price = $rawPrice;
             $api_price = fn_novoton_holidays_get_api()->pricing()->applyCommission($rawPrice);
